- adjust package name of icon addon to fix detekt failure
