--- simple-icons-typeface-library/src/main/java/com/mikepenz/iconics/typeface/library/simple_icons/SimpleIcons.kt
@@ -1,2329 +1,0 @@
-/*
- * Copyright 2022 Mike Penz
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.mikepenz.iconics.typeface.library.simple_icons
-
-import com.mikepenz.iconics.typeface.IIcon
-import com.mikepenz.iconics.typeface.ITypeface
-import java.util.LinkedList
-
-@Suppress("EnumEntryName")
-object SimpleIcons : ITypeface {
-
-    override val fontRes: Int
-        get() = R.font.simple_icons_font_v17_0_0
-
-    override val characters: Map<String, Char> by lazy {
-        Icon.values().associate { it.name to it.character }
-    }
-    
-    override val mappingPrefix: String
-        get() = "sim"
-
-    override val fontName: String
-        get() = "Simple-Icons"
-
-    override val version: String
-        get() = "17.0.0"
-
-    override val iconCount: Int
-        get() = characters.size
-
-    override val icons: List<String>
-        get() = characters.keys.toCollection(LinkedList())
-
-    override val author: String
-        get() = "Simple-Icons"
-
-    override val url: String
-        get() = "https://simpleicons.org/"
-
-    override val description: String
-        get() = "Icons for Popular Brands"
-
-    override val license: String
-        get() = "CC0 1.0 Universal"
-
-    override val licenseUrl: String
-        get() = "https://github.com/simple-icons/simple-icons/blob/develop/LICENSE.md"
-
-    override fun getIcon(key: String): IIcon = Icon.valueOf(key)
-
-    enum class Icon constructor(override val character: Char) : IIcon {
-        sim_42('\ue904'),
-		sim_abbott('\ue907'),
-		sim_abbrobotstudio('\ue908'),
-		sim_abbvie('\ue909'),
-		sim_abletonlive('\ue90a'),
-		sim_aboutdotme('\ue90b'),
-		sim_abstract('\ue90c'),
-		sim_academia('\ue90d'),
-		sim_accenture('\ue90e'),
-		sim_acclaim('\ue90f'),
-		sim_accusoft('\ue910'),
-		sim_acer('\ue911'),
-		sim_acm('\ue912'),
-		sim_actigraph('\ue913'),
-		sim_activision('\ue914'),
-		sim_adafruit('\ue915'),
-		sim_adblock('\ue916'),
-		sim_adblockplus('\ue917'),
-		sim_addthis('\ue918'),
-		sim_adguard('\ue919'),
-		sim_adidas('\ue91a'),
-		sim_adobe('\ue91b'),
-		sim_adobeacrobatreader('\ue91c'),
-		sim_adobeaftereffects('\ue91d'),
-		sim_adobeaudition('\ue91e'),
-		sim_adobecreativecloud('\ue91f'),
-		sim_adobedreamweaver('\ue920'),
-		sim_adobefonts('\ue921'),
-		sim_adobeillustrator('\ue922'),
-		sim_adobeindesign('\ue923'),
-		sim_adobelightroom('\ue924'),
-		sim_adobelightroomclassic('\ue925'),
-		sim_adobephotoshop('\ue926'),
-		sim_adobepremierepro('\ue927'),
-		sim_adobexd('\ue928'),
-		sim_adonisjs('\ue929'),
-		sim_adyen('\ue92a'),
-		sim_aerlingus('\ue92b'),
-		sim_aeroflot('\ue92c'),
-		sim_aeromexico('\ue92d'),
-		sim_aerospike('\ue92e'),
-		sim_aew('\ue92f'),
-		sim_affinity('\ue930'),
-		sim_affinitydesigner('\ue931'),
-		sim_affinityphoto('\ue932'),
-		sim_affinitypublisher('\ue933'),
-		sim_aframe('\ue934'),
-		sim_aib('\ue935'),
-		sim_aidungeon('\ue936'),
-		sim_aiohttp('\ue937'),
-		sim_aiqfome('\ue938'),
-		sim_airasia('\ue939'),
-		sim_airbnb('\ue93a'),
-		sim_airbus('\ue93b'),
-		sim_aircall('\ue93c'),
-		sim_aircanada('\ue93d'),
-		sim_airchina('\ue93e'),
-		sim_airfrance('\ue93f'),
-		sim_airplayaudio('\ue940'),
-		sim_airplayvideo('\ue941'),
-		sim_airtable('\ue942'),
-		sim_ajv('\ue943'),
-		sim_alacritty('\ue944'),
-		sim_albertheijn('\ue945'),
-		sim_aldinord('\ue946'),
-		sim_aldisud('\ue947'),
-		sim_alfaromeo('\ue948'),
-		sim_alfred('\ue949'),
-		sim_algolia('\ue94a'),
-		sim_alibabacloud('\ue94b'),
-		sim_alibabadotcom('\ue94c'),
-		sim_aliexpress('\ue94d'),
-		sim_alipay('\ue94e'),
-		sim_alitalia('\ue94f'),
-		sim_allegro('\ue950'),
-		sim_alliedmodders('\ue951'),
-		sim_allocine('\ue952'),
-		sim_alltrails('\ue953'),
-		sim_alpinedotjs('\ue954'),
-		sim_alpinelinux('\ue955'),
-		sim_altiumdesigner('\ue956'),
-		sim_amazon('\ue957'),
-		sim_amazonalexa('\ue958'),
-		sim_amazonaws('\ue959'),
-		sim_amazondynamodb('\ue95a'),
-		sim_amazonfiretv('\ue95b'),
-		sim_amazonlumberyard('\ue95c'),
-		sim_amazonpay('\ue95d'),
-		sim_amazonprime('\ue95e'),
-		sim_amazons3('\ue95f'),
-		sim_amd('\ue960'),
-		sim_americanairlines('\ue961'),
-		sim_americanexpress('\ue962'),
-		sim_amg('\ue963'),
-		sim_amp('\ue964'),
-		sim_amul('\ue965'),
-		sim_ana('\ue966'),
-		sim_anaconda('\ue967'),
-		sim_analogue('\ue968'),
-		sim_anchor('\ue969'),
-		sim_andela('\ue96a'),
-		sim_android('\ue96b'),
-		sim_androidauto('\ue96c'),
-		sim_androidstudio('\ue96d'),
-		sim_angellist('\ue96e'),
-		sim_angular('\ue96f'),
-		sim_angularjs('\ue970'),
-		sim_angularuniversal('\ue971'),
-		sim_anilist('\ue972'),
-		sim_ansible('\ue973'),
-		sim_ansys('\ue974'),
-		sim_anta('\ue975'),
-		sim_antdesign('\ue976'),
-		sim_antena3('\ue977'),
-		sim_anydesk('\ue978'),
-		sim_aol('\ue979'),
-		sim_apache('\ue97a'),
-		sim_apacheairflow('\ue97b'),
-		sim_apacheant('\ue97c'),
-		sim_apachecassandra('\ue97d'),
-		sim_apachecloudstack('\ue97e'),
-		sim_apachecordova('\ue97f'),
-		sim_apachecouchdb('\ue980'),
-		sim_apachedruid('\ue981'),
-		sim_apacheecharts('\ue982'),
-		sim_apacheflink('\ue983'),
-		sim_apachegroovy('\ue984'),
-		sim_apachehadoop('\ue985'),
-		sim_apachehive('\ue986'),
-		sim_apachejmeter('\ue987'),
-		sim_apachekafka('\ue988'),
-		sim_apachekylin('\ue989'),
-		sim_apachemaven('\ue98a'),
-		sim_apachenetbeanside('\ue98b'),
-		sim_apacheopenoffice('\ue98c'),
-		sim_apachepulsar('\ue98d'),
-		sim_apacherocketmq('\ue98e'),
-		sim_apachesolr('\ue98f'),
-		sim_apachespark('\ue990'),
-		sim_apachetomcat('\ue991'),
-		sim_aparat('\ue992'),
-		sim_apollographql('\ue993'),
-		sim_apostrophe('\ue994'),
-		sim_appannie('\ue995'),
-		sim_appian('\ue996'),
-		sim_apple('\ue997'),
-		sim_applearcade('\ue998'),
-		sim_applemusic('\ue999'),
-		sim_applepay('\ue99a'),
-		sim_applepodcasts('\ue99b'),
-		sim_appletv('\ue99c'),
-		sim_appsignal('\ue99d'),
-		sim_appstore('\ue99e'),
-		sim_appveyor('\ue99f'),
-		sim_appwrite('\ue9a0'),
-		sim_aqua('\ue9a1'),
-		sim_aral('\ue9a2'),
-		sim_arangodb('\ue9a3'),
-		sim_archicad('\ue9a4'),
-		sim_archiveofourown('\ue9a5'),
-		sim_archlinux('\ue9a6'),
-		sim_ardour('\ue9a7'),
-		sim_arduino('\ue9a8'),
-		sim_argos('\ue9a9'),
-		sim_arkecosystem('\ue9aa'),
-		sim_arlo('\ue9ab'),
-		sim_arm('\ue9ac'),
-		sim_artifacthub('\ue9ad'),
-		sim_artixlinux('\ue9ae'),
-		sim_artstation('\ue9af'),
-		sim_arxiv('\ue9b0'),
-		sim_asana('\ue9b1'),
-		sim_asciidoctor('\ue9b2'),
-		sim_asciinema('\ue9b3'),
-		sim_asda('\ue9b4'),
-		sim_aseprite('\ue9b5'),
-		sim_askfm('\ue9b6'),
-		sim_askubuntu('\ue9b7'),
-		sim_assemblyscript('\ue9b8'),
-		sim_astonmartin('\ue9b9'),
-		sim_astro('\ue9ba'),
-		sim_asus('\ue9bb'),
-		sim_atandt('\ue9bc'),
-		sim_atari('\ue9bd'),
-		sim_atlassian('\ue9be'),
-		sim_atom('\ue9bf'),
-		sim_auchan('\ue9c0'),
-		sim_audacity('\ue9c1'),
-		sim_audi('\ue9c2'),
-		sim_audible('\ue9c3'),
-		sim_audioboom('\ue9c4'),
-		sim_audiomack('\ue9c5'),
-		sim_audiotechnica('\ue9c6'),
-		sim_aurelia('\ue9c7'),
-		sim_auth0('\ue9c8'),
-		sim_authy('\ue9c9'),
-		sim_autodesk('\ue9ca'),
-		sim_autohotkey('\ue9cb'),
-		sim_automattic('\ue9cc'),
-		sim_autoprefixer('\ue9cd'),
-		sim_avast('\ue9ce'),
-		sim_awesomelists('\ue9cf'),
-		sim_awesomewm('\ue9d0'),
-		sim_awsamplify('\ue9d1'),
-		sim_awslambda('\ue9d2'),
-		sim_azureartifacts('\ue9d3'),
-		sim_azuredataexplorer('\ue9d4'),
-		sim_azuredevops('\ue9d5'),
-		sim_azurefunctions('\ue9d6'),
-		sim_azurepipelines('\ue9d7'),
-		sim_babel('\ue9d8'),
-		sim_backbonedotjs('\ue9d9'),
-		sim_badgr('\ue9da'),
-		sim_badoo('\ue9db'),
-		sim_baidu('\ue9dc'),
-		sim_bamboo('\ue9dd'),
-		sim_bandcamp('\ue9de'),
-		sim_bandlab('\ue9df'),
-		sim_bandrautomation('\ue9e0'),
-		sim_bandsintown('\ue9e1'),
-		sim_bankofamerica('\ue9e2'),
-		sim_barclays('\ue9e3'),
-		sim_baremetrics('\ue9e4'),
-		sim_basecamp('\ue9e5'),
-		sim_bata('\ue9e6'),
-		sim_bathasu('\ue9e7'),
-		sim_battledotnet('\ue9e8'),
-		sim_bbc('\ue9e9'),
-		sim_bbciplayer('\ue9ea'),
-		sim_beatport('\ue9eb'),
-		sim_beats('\ue9ec'),
-		sim_beatsbydre('\ue9ed'),
-		sim_behance('\ue9ee'),
-		sim_beijingsubway('\ue9ef'),
-		sim_bem('\ue9f0'),
-		sim_bentley('\ue9f1'),
-		sim_betfair('\ue9f2'),
-		sim_bigbasket('\ue9f3'),
-		sim_bigbluebutton('\ue9f4'),
-		sim_bigcartel('\ue9f5'),
-		sim_bigcommerce('\ue9f6'),
-		sim_bilibili('\ue9f7'),
-		sim_binance('\ue9f8'),
-		sim_biolink('\ue9f9'),
-		sim_bit('\ue9fa'),
-		sim_bitbucket('\ue9fb'),
-		sim_bitcoin('\ue9fc'),
-		sim_bitcoincash('\ue9fd'),
-		sim_bitcoinsv('\ue9fe'),
-		sim_bitdefender('\ue9ff'),
-		sim_bitly('\uea00'),
-		sim_bitrise('\uea01'),
-		sim_bitwarden('\uea02'),
-		sim_bitwig('\uea03'),
-		sim_blackberry('\uea04'),
-		sim_blazemeter('\uea05'),
-		sim_blazor('\uea06'),
-		sim_blender('\uea07'),
-		sim_blockchaindotcom('\uea08'),
-		sim_blogger('\uea09'),
-		sim_bloglovin('\uea0a'),
-		sim_blueprint('\uea0b'),
-		sim_bluetooth('\uea0c'),
-		sim_bmcsoftware('\uea0d'),
-		sim_bmw('\uea0e'),
-		sim_boehringeringelheim('\uea0f'),
-		sim_boeing('\uea10'),
-		sim_bookbub('\uea11'),
-		sim_bookmeter('\uea12'),
-		sim_bookstack('\uea13'),
-		sim_boost('\uea14'),
-		sim_boots('\uea15'),
-		sim_bootstrap('\uea16'),
-		sim_bosch('\uea17'),
-		sim_bose('\uea18'),
-		sim_boulanger('\uea19'),
-		sim_bower('\uea1a'),
-		sim_box('\uea1b'),
-		sim_boxysvg('\uea1c'),
-		sim_brandfolder('\uea1d'),
-		sim_brave('\uea1e'),
-		sim_breaker('\uea1f'),
-		sim_britishairways('\uea20'),
-		sim_broadcom('\uea21'),
-		sim_bt('\uea22'),
-		sim_buddy('\uea23'),
-		sim_budibase('\uea24'),
-		sim_buefy('\uea25'),
-		sim_buffer('\uea26'),
-		sim_bugatti('\uea27'),
-		sim_bugcrowd('\uea28'),
-		sim_bugsnag('\uea29'),
-		sim_buildkite('\uea2a'),
-		sim_bukalapak('\uea2b'),
-		sim_bulma('\uea2c'),
-		sim_bunq('\uea2d'),
-		sim_burgerking('\uea2e'),
-		sim_buymeacoffee('\uea2f'),
-		sim_buzzfeed('\uea30'),
-		sim_byjus('\uea31'),
-		sim_byte('\uea32'),
-		sim_bytedance('\uea33'),
-		sim_c('\uea34'),
-		sim_cachet('\uea35'),
-		sim_caffeine('\uea36'),
-		sim_cairometro('\uea37'),
-		sim_cakephp('\uea38'),
-		sim_campaignmonitor('\uea39'),
-		sim_canonical('\uea3a'),
-		sim_canva('\uea3b'),
-		sim_capacitor('\uea3c'),
-		sim_carrefour('\uea3d'),
-		sim_carthrottle('\uea3e'),
-		sim_carto('\uea3f'),
-		sim_cashapp('\uea40'),
-		sim_castbox('\uea41'),
-		sim_castorama('\uea42'),
-		sim_castro('\uea43'),
-		sim_caterpillar('\uea44'),
-		sim_cbs('\uea45'),
-		sim_cdprojekt('\uea46'),
-		sim_celery('\uea47'),
-		sim_centos('\uea48'),
-		sim_ceph('\uea49'),
-		sim_cesium('\uea4a'),
-		sim_chai('\uea4b'),
-		sim_chainlink('\uea4c'),
-		sim_chakraui('\uea4d'),
-		sim_chan('\ue902'),
-		sim_chartdotjs('\uea4e'),
-		sim_chartmogul('\uea4f'),
-		sim_chase('\uea50'),
-		sim_chatbot('\uea51'),
-		sim_checkio('\uea52'),
-		sim_checkmarx('\uea53'),
-		sim_chef('\uea54'),
-		sim_chemex('\uea55'),
-		sim_chevrolet('\uea56'),
-		sim_chinaeasternairlines('\uea57'),
-		sim_chinasouthernairlines('\uea58'),
-		sim_chocolatey('\uea59'),
-		sim_chromecast('\uea5a'),
-		sim_chrysler('\uea5b'),
-		sim_chupachups('\uea5c'),
-		sim_cilium('\uea5d'),
-		sim_cinema4d('\uea5e'),
-		sim_circle('\uea5f'),
-		sim_circleci('\uea60'),
-		sim_cirrusci('\uea61'),
-		sim_cisco('\uea62'),
-		sim_citrix('\uea63'),
-		sim_citroen('\uea64'),
-		sim_civicrm('\uea65'),
-		sim_civo('\uea66'),
-		sim_ckeditor4('\uea67'),
-		sim_claris('\uea68'),
-		sim_clickhouse('\uea69'),
-		sim_clickup('\uea6a'),
-		sim_clion('\uea6b'),
-		sim_cliqz('\uea6c'),
-		sim_clockify('\uea6d'),
-		sim_clojure('\uea6e'),
-		sim_cloud66('\uea6f'),
-		sim_cloudbees('\uea70'),
-		sim_cloudcannon('\uea71'),
-		sim_cloudera('\uea72'),
-		sim_cloudflare('\uea73'),
-		sim_cloudfoundry('\uea74'),
-		sim_cloudsmith('\uea75'),
-		sim_cloudways('\uea76'),
-		sim_clubhouse('\uea77'),
-		sim_clyp('\uea78'),
-		sim_cmake('\uea79'),
-		sim_cncf('\uea7a'),
-		sim_cnn('\uea7b'),
-		sim_cockpit('\uea7c'),
-		sim_cockroachlabs('\uea7d'),
-		sim_cocoapods('\uea7e'),
-		sim_cocos('\uea7f'),
-		sim_coda('\uea80'),
-		sim_codacy('\uea81'),
-		sim_codeberg('\uea82'),
-		sim_codecademy('\uea83'),
-		sim_codeceptjs('\uea84'),
-		sim_codechef('\uea85'),
-		sim_codeclimate('\uea86'),
-		sim_codecov('\uea87'),
-		sim_codefactor('\uea88'),
-		sim_codeforces('\uea89'),
-		sim_codeigniter('\uea8a'),
-		sim_codemagic('\uea8b'),
-		sim_codemirror('\uea8c'),
-		sim_codenewbie('\uea8d'),
-		sim_codepen('\uea8e'),
-		sim_codeproject('\uea8f'),
-		sim_codereview('\uea90'),
-		sim_codersrank('\uea91'),
-		sim_coderwall('\uea92'),
-		sim_codesandbox('\uea93'),
-		sim_codeship('\uea94'),
-		sim_codewars('\uea95'),
-		sim_codingame('\uea96'),
-		sim_codingninjas('\uea97'),
-		sim_codio('\uea98'),
-		sim_coffeescript('\uea99'),
-		sim_cognizant('\uea9a'),
-		sim_coil('\uea9b'),
-		sim_coinbase('\uea9c'),
-		sim_coinmarketcap('\uea9d'),
-		sim_commerzbank('\uea9e'),
-		sim_commonworkflowlanguage('\uea9f'),
-		sim_composer('\ueaa0'),
-		sim_comsol('\ueaa1'),
-		sim_conan('\ueaa2'),
-		sim_concourse('\ueaa3'),
-		sim_condaforge('\ueaa4'),
-		sim_conekta('\ueaa5'),
-		sim_confluence('\ueaa6'),
-		sim_construct3('\ueaa7'),
-		sim_consul('\ueaa8'),
-		sim_contactlesspayment('\ueaa9'),
-		sim_containerd('\ueaaa'),
-		sim_contentful('\ueaab'),
-		sim_conventionalcommits('\ueaac'),
-		sim_convertio('\ueaad'),
-		sim_cookiecutter('\ueaae'),
-		sim_coop('\ueaaf'),
-		sim_cora('\ueab0'),
-		sim_coronaengine('\ueab1'),
-		sim_coronarenderer('\ueab2'),
-		sim_corsair('\ueab3'),
-		sim_couchbase('\ueab4'),
-		sim_counterstrike('\ueab5'),
-		sim_countingworkspro('\ueab6'),
-		sim_coursera('\ueab7'),
-		sim_coveralls('\ueab8'),
-		sim_cpanel('\ueab9'),
-		sim_cplusplus('\ueaba'),
-		sim_craftcms('\ueabb'),
-		sim_cratedb('\ueabc'),
-		sim_createreactapp('\ueabd'),
-		sim_creativecommons('\ueabe'),
-		sim_credly('\ueabf'),
-		sim_crehana('\ueac0'),
-		sim_crowdin('\ueac1'),
-		sim_crowdsource('\ueac2'),
-		sim_crunchbase('\ueac3'),
-		sim_crunchyroll('\ueac4'),
-		sim_cryengine('\ueac5'),
-		sim_crystal('\ueac6'),
-		sim_csharp('\ueac7'),
-		sim_css3('\ueac8'),
-		sim_cssmodules('\ueac9'),
-		sim_csswizardry('\ueaca'),
-		sim_cucumber('\ueacb'),
-		sim_curl('\ueacc'),
-		sim_curseforge('\ueacd'),
-		sim_cycling74('\ueace'),
-		sim_cypress('\ueacf'),
-		sim_cytoscapedotjs('\uead0'),
-		sim_d('\ue903'),
-		sim_d3dotjs('\uead1'),
-		sim_dacia('\uead2'),
-		sim_daf('\uead3'),
-		sim_dailymotion('\uead4'),
-		sim_daimler('\uead5'),
-		sim_darkreader('\uead6'),
-		sim_dart('\uead7'),
-		sim_darty('\uead8'),
-		sim_daserste('\uead9'),
-		sim_dash('\ueada'),
-		sim_dashlane('\ueadb'),
-		sim_dask('\ueadc'),
-		sim_dassaultsystemes('\ueadd'),
-		sim_databricks('\ueade'),
-		sim_datacamp('\ueadf'),
-		sim_datadog('\ueae0'),
-		sim_datadotai('\ueae1'),
-		sim_datagrip('\ueae2'),
-		sim_dataiku('\ueae3'),
-		sim_datastax('\ueae4'),
-		sim_dataversioncontrol('\ueae5'),
-		sim_datocms('\ueae6'),
-		sim_datto('\ueae7'),
-		sim_dazn('\ueae8'),
-		sim_dblp('\ueae9'),
-		sim_dbt('\ueaea'),
-		sim_dcentertainment('\ueaeb'),
-		sim_debian('\ueaec'),
-		sim_dedge('\ueaed'),
-		sim_deepin('\ueaee'),
-		sim_deepnote('\ueaef'),
-		sim_deezer('\ueaf0'),
-		sim_delicious('\ueaf1'),
-		sim_deliveroo('\ueaf2'),
-		sim_dell('\ueaf3'),
-		sim_delonghi('\ueaf4'),
-		sim_delphi('\ueaf5'),
-		sim_delta('\ueaf6'),
-		sim_deno('\ueaf7'),
-		sim_dependabot('\ueaf8'),
-		sim_derspiegel('\ueaf9'),
-		sim_designernews('\ueafa'),
-		sim_deutschebahn('\ueafb'),
-		sim_deutschebank('\ueafc'),
-		sim_devdotto('\ueafd'),
-		sim_devexpress('\ueafe'),
-		sim_deviantart('\ueaff'),
-		sim_devpost('\ueb00'),
-		sim_devrant('\ueb01'),
-		sim_dgraph('\ueb02'),
-		sim_dhl('\ueb03'),
-		sim_diagramsdotnet('\ueb04'),
-		sim_dialogflow('\ueb05'),
-		sim_diaspora('\ueb06'),
-		sim_digg('\ueb07'),
-		sim_digikeyelectronics('\ueb08'),
-		sim_digitalocean('\ueb09'),
-		sim_dior('\ueb0a'),
-		sim_directus('\ueb0b'),
-		sim_discogs('\ueb0c'),
-		sim_discord('\ueb0d'),
-		sim_discourse('\ueb0e'),
-		sim_discover('\ueb0f'),
-		sim_disqus('\ueb10'),
-		sim_disroot('\ueb11'),
-		sim_django('\ueb12'),
-		sim_dlib('\ueb13'),
-		sim_dlna('\ueb14'),
-		sim_dm('\ueb15'),
-		sim_docker('\ueb16'),
-		sim_docsdotrs('\ueb17'),
-		sim_docusign('\ueb18'),
-		sim_dogecoin('\ueb19'),
-		sim_dolby('\ueb1a'),
-		sim_doordash('\ueb1b'),
-		sim_dotnet('\ueb1c'),
-		sim_douban('\ueb1d'),
-		sim_doubanread('\ueb1e'),
-		sim_dpd('\ueb1f'),
-		sim_dragonframe('\ueb20'),
-		sim_draugiemdotlv('\ueb21'),
-		sim_dribbble('\ueb22'),
-		sim_drone('\ueb23'),
-		sim_drooble('\ueb24'),
-		sim_dropbox('\ueb25'),
-		sim_drupal('\ueb26'),
-		sim_dsautomobiles('\ueb27'),
-		sim_dtube('\ueb28'),
-		sim_duckduckgo('\ueb29'),
-		sim_dungeonsanddragons('\ueb2a'),
-		sim_dunked('\ueb2b'),
-		sim_duolingo('\ueb2c'),
-		sim_dvc('\ueb2d'),
-		sim_dwavesystems('\ueb2e'),
-		sim_dwm('\ueb2f'),
-		sim_dynamics365('\ueb30'),
-		sim_dynatrace('\ueb31'),
-		sim_e('\ueb32'),
-		sim_ea('\ueb33'),
-		sim_eagle('\ueb34'),
-		sim_easyjet('\ueb35'),
-		sim_ebay('\ueb36'),
-		sim_eclipseche('\ueb37'),
-		sim_eclipseide('\ueb38'),
-		sim_eclipsejetty('\ueb39'),
-		sim_eclipsemosquitto('\ueb3a'),
-		sim_eclipsevertdotx('\ueb3b'),
-		sim_edeka('\ueb3c'),
-		sim_editorconfig('\ueb3d'),
-		sim_edotleclerc('\ueb3e'),
-		sim_edx('\ueb3f'),
-		sim_egghead('\ueb40'),
-		sim_egnyte('\ueb41'),
-		sim_eightsleep('\ueb42'),
-		sim_elastic('\ueb43'),
-		sim_elasticcloud('\ueb44'),
-		sim_elasticsearch('\ueb45'),
-		sim_elasticstack('\ueb46'),
-		sim_electron('\ueb47'),
-		sim_element('\ueb48'),
-		sim_elementary('\ueb49'),
-		sim_eleventy('\ueb4a'),
-		sim_elixir('\ueb4b'),
-		sim_eljueves('\ueb4c'),
-		sim_ello('\ueb4d'),
-		sim_elm('\ueb4e'),
-		sim_elsevier('\ueb4f'),
-		sim_embarcadero('\ueb50'),
-		sim_emberdotjs('\ueb51'),
-		sim_emby('\ueb52'),
-		sim_emirates('\ueb53'),
-		sim_emlakjet('\ueb54'),
-		sim_empirekred('\ueb55'),
-		sim_enpass('\ueb56'),
-		sim_enterprisedb('\ueb57'),
-		sim_envato('\ueb58'),
-		sim_epel('\ueb59'),
-		sim_epicgames('\ueb5a'),
-		sim_epson('\ueb5b'),
-		sim_equinixmetal('\ueb5c'),
-		sim_erlang('\ueb5d'),
-		sim_esbuild('\ueb5e'),
-		sim_esea('\ueb5f'),
-		sim_eslgaming('\ueb60'),
-		sim_eslint('\ueb61'),
-		sim_esphome('\ueb62'),
-		sim_espressif('\ueb63'),
-		sim_etcd('\ueb64'),
-		sim_ethereum('\ueb65'),
-		sim_ethiopianairlines('\ueb66'),
-		sim_etihadairways('\ueb67'),
-		sim_etsy('\ueb68'),
-		sim_eventbrite('\ueb69'),
-		sim_eventstore('\ueb6a'),
-		sim_evernote('\ueb6b'),
-		sim_exercism('\ueb6c'),
-		sim_exordo('\ueb6d'),
-		sim_exoscale('\ueb6e'),
-		sim_expensify('\ueb6f'),
-		sim_expertsexchange('\ueb70'),
-		sim_expo('\ueb71'),
-		sim_express('\ueb72'),
-		sim_expressvpn('\ueb73'),
-		sim_eyeem('\ueb74'),
-		sim_facebook('\ueb75'),
-		sim_facebookgaming('\ueb76'),
-		sim_facebooklive('\ueb77'),
-		sim_faceit('\ueb78'),
-		sim_facepunch('\ueb79'),
-		sim_falcon('\ueb7a'),
-		sim_fampay('\ueb7b'),
-		sim_fandango('\ueb7c'),
-		sim_fandom('\ueb7d'),
-		sim_farfetch('\ueb7e'),
-		sim_fastapi('\ueb7f'),
-		sim_fastify('\ueb80'),
-		sim_fastlane('\ueb81'),
-		sim_fastly('\ueb82'),
-		sim_fathom('\ueb83'),
-		sim_favro('\ueb84'),
-		sim_fdroid('\ueb85'),
-		sim_feathub('\ueb86'),
-		sim_fedex('\ueb87'),
-		sim_fedora('\ueb88'),
-		sim_feedly('\ueb89'),
-		sim_ferrari('\ueb8a'),
-		sim_ferrarinv('\ueb8b'),
-		sim_ffmpeg('\ueb8c'),
-		sim_fiat('\ueb8d'),
-		sim_fidoalliance('\ueb8e'),
-		sim_fifa('\ueb8f'),
-		sim_figma('\ueb90'),
-		sim_figshare('\ueb91'),
-		sim_fila('\ueb92'),
-		sim_files('\ueb93'),
-		sim_filezilla('\ueb94'),
-		sim_fing('\ueb95'),
-		sim_firebase('\ueb96'),
-		sim_firefox('\ueb97'),
-		sim_firefoxbrowser('\ueb98'),
-		sim_first('\ueb99'),
-		sim_fitbit('\ueb9a'),
-		sim_fite('\ueb9b'),
-		sim_fivem('\ueb9c'),
-		sim_fiverr('\ueb9d'),
-		sim_flask('\ueb9e'),
-		sim_flat('\ueb9f'),
-		sim_flathub('\ueba0'),
-		sim_flattr('\ueba1'),
-		sim_flickr('\ueba2'),
-		sim_flipboard('\ueba3'),
-		sim_flipkart('\ueba4'),
-		sim_floatplane('\ueba5'),
-		sim_flood('\ueba6'),
-		sim_fluentbit('\ueba7'),
-		sim_fluentd('\ueba8'),
-		sim_flutter('\ueba9'),
-		sim_flyway('\uebaa'),
-		sim_fmod('\uebab'),
-		sim_fnac('\uebac'),
-		sim_folium('\uebad'),
-		sim_fonoma('\uebae'),
-		sim_fontawesome('\uebaf'),
-		sim_fontbase('\uebb0'),
-		sim_foodpanda('\uebb1'),
-		sim_ford('\uebb2'),
-		sim_forestry('\uebb3'),
-		sim_formstack('\uebb4'),
-		sim_fortinet('\uebb5'),
-		sim_fortran('\uebb6'),
-		sim_fossa('\uebb7'),
-		sim_fossilscm('\uebb8'),
-		sim_foursquare('\uebb9'),
-		sim_foursquarecityguide('\uebba'),
-		sim_foxtel('\uebbb'),
-		sim_fozzy('\uebbc'),
-		sim_framer('\uebbd'),
-		sim_framework7('\uebbe'),
-		sim_franprix('\uebbf'),
-		sim_fraunhofergesellschaft('\uebc0'),
-		sim_freebsd('\uebc1'),
-		sim_freecodecamp('\uebc2'),
-		sim_freedesktopdotorg('\uebc3'),
-		sim_freelancer('\uebc4'),
-		sim_freenas('\uebc5'),
-		sim_frontendmentor('\uebc6'),
-		sim_fsecure('\uebc7'),
-		sim_fujifilm('\uebc8'),
-		sim_fujitsu('\uebc9'),
-		sim_funimation('\uebca'),
-		sim_furaffinity('\uebcb'),
-		sim_furrynetwork('\uebcc'),
-		sim_futurelearn('\uebcd'),
-		sim_g2('\uebce'),
-		sim_g2a('\uebcf'),
-		sim_gameandwatch('\uebd0'),
-		sim_gamejolt('\uebd1'),
-		sim_garmin('\uebd2'),
-		sim_gatling('\uebd3'),
-		sim_gatsby('\uebd4'),
-		sim_geant('\uebd5'),
-		sim_geeksforgeeks('\uebd6'),
-		sim_generalelectric('\uebd7'),
-		sim_generalmotors('\uebd8'),
-		sim_genius('\uebd9'),
-		sim_gentoo('\uebda'),
-		sim_geocaching('\uebdb'),
-		sim_gerrit('\uebdc'),
-		sim_ghost('\uebdd'),
-		sim_ghostery('\uebde'),
-		sim_gimp('\uebdf'),
-		sim_giphy('\uebe0'),
-		sim_git('\uebe1'),
-		sim_gitbook('\uebe2'),
-		sim_gitea('\uebe3'),
-		sim_gitee('\uebe4'),
-		sim_gitextensions('\uebe5'),
-		sim_github('\uebe6'),
-		sim_githubactions('\uebe7'),
-		sim_githubpages('\uebe8'),
-		sim_githubsponsors('\uebe9'),
-		sim_gitignoredotio('\uebea'),
-		sim_gitkraken('\uebeb'),
-		sim_gitlab('\uebec'),
-		sim_gitlfs('\uebed'),
-		sim_gitpod('\uebee'),
-		sim_gitter('\uebef'),
-		sim_glassdoor('\uebf0'),
-		sim_glitch('\uebf1'),
-		sim_globus('\uebf2'),
-		sim_gmail('\uebf3'),
-		sim_gnome('\uebf4'),
-		sim_gnometerminal('\uebf5'),
-		sim_gnu('\uebf6'),
-		sim_gnubash('\uebf7'),
-		sim_gnuemacs('\uebf8'),
-		sim_gnuicecat('\uebf9'),
-		sim_gnuprivacyguard('\uebfa'),
-		sim_gnusocial('\uebfb'),
-		sim_go('\uebfc'),
-		sim_gocd('\uebfd'),
-		sim_godaddy('\uebfe'),
-		sim_godotengine('\uebff'),
-		sim_gofundme('\uec00'),
-		sim_gogdotcom('\uec01'),
-		sim_goland('\uec02'),
-		sim_goldenline('\uec03'),
-		sim_goodreads('\uec04'),
-		sim_google('\uec05'),
-		sim_googleadmob('\uec06'),
-		sim_googleads('\uec07'),
-		sim_googleadsense('\uec08'),
-		sim_googleanalytics('\uec09'),
-		sim_googleassistant('\uec0a'),
-		sim_googlecalendar('\uec0b'),
-		sim_googlecardboard('\uec0c'),
-		sim_googlechat('\uec0d'),
-		sim_googlechrome('\uec0e'),
-		sim_googleclassroom('\uec0f'),
-		sim_googlecloud('\uec10'),
-		sim_googlecolab('\uec11'),
-		sim_googledomains('\uec12'),
-		sim_googledrive('\uec13'),
-		sim_googleearth('\uec14'),
-		sim_googlefit('\uec15'),
-		sim_googlefonts('\uec16'),
-		sim_googlehangouts('\uec17'),
-		sim_googlekeep('\uec18'),
-		sim_googlelens('\uec19'),
-		sim_googlemaps('\uec1a'),
-		sim_googlemarketingplatform('\uec1b'),
-		sim_googlemeet('\uec1c'),
-		sim_googlemessages('\uec1d'),
-		sim_googlemybusiness('\uec1e'),
-		sim_googlenearby('\uec1f'),
-		sim_googlenews('\uec20'),
-		sim_googleoptimize('\uec21'),
-		sim_googlepay('\uec22'),
-		sim_googlephotos('\uec23'),
-		sim_googleplay('\uec24'),
-		sim_googlepodcasts('\uec25'),
-		sim_googlescholar('\uec26'),
-		sim_googlesearchconsole('\uec27'),
-		sim_googlesheets('\uec28'),
-		sim_googlestreetview('\uec29'),
-		sim_googletagmanager('\uec2a'),
-		sim_googletranslate('\uec2b'),
-		sim_gotomeeting('\uec2c'),
-		sim_grab('\uec2d'),
-		sim_gradle('\uec2e'),
-		sim_grafana('\uec2f'),
-		sim_grammarly('\uec30'),
-		sim_grandfrais('\uec31'),
-		sim_graphql('\uec32'),
-		sim_grav('\uec33'),
-		sim_gravatar('\uec34'),
-		sim_graylog('\uec35'),
-		sim_greensock('\uec36'),
-		sim_griddotai('\uec37'),
-		sim_gridsome('\uec38'),
-		sim_groupme('\uec39'),
-		sim_groupon('\uec3a'),
-		sim_grubhub('\uec3b'),
-		sim_grunt('\uec3c'),
-		sim_gtk('\uec3d'),
-		sim_guangzhoumetro('\uec3e'),
-		sim_guilded('\uec3f'),
-		sim_gulp('\uec40'),
-		sim_gumroad('\uec41'),
-		sim_gumtree('\uec42'),
-		sim_gunicorn('\uec43'),
-		sim_gurobi('\uec44'),
-		sim_gutenberg('\uec45'),
-		sim_habr('\uec46'),
-		sim_hackaday('\uec47'),
-		sim_hackclub('\uec48'),
-		sim_hackerearth('\uec49'),
-		sim_hackernoon('\uec4a'),
-		sim_hackerone('\uec4b'),
-		sim_hackerrank('\uec4c'),
-		sim_hackster('\uec4d'),
-		sim_hackthebox('\uec4e'),
-		sim_handlebarsdotjs('\uec4f'),
-		sim_handshake('\uec50'),
-		sim_handshake_protocol('\uec51'),
-		sim_happycow('\uec52'),
-		sim_harbor('\uec53'),
-		sim_hashnode('\uec54'),
-		sim_haskell('\uec55'),
-		sim_hasura('\uec56'),
-		sim_hatenabookmark('\uec57'),
-		sim_haveibeenpwned('\uec58'),
-		sim_haxe('\uec59'),
-		sim_hbo('\uec5a'),
-		sim_hcl('\uec5b'),
-		sim_headlessui('\uec5c'),
-		sim_headspace('\uec5d'),
-		sim_hellofresh('\uec5e'),
-		sim_hellyhansen('\uec5f'),
-		sim_helm('\uec60'),
-		sim_helpdesk('\uec61'),
-		sim_helpscout('\uec62'),
-		sim_here('\uec63'),
-		sim_heroku('\uec64'),
-		sim_hetzner('\uec65'),
-		sim_hexo('\uec66'),
-		sim_hey('\uec67'),
-		sim_hibernate('\uec68'),
-		sim_hilton('\uec69'),
-		sim_hitachi('\uec6a'),
-		sim_hive('\uec6b'),
-		sim_hive_blockchain('\uec6c'),
-		sim_homeadvisor('\uec6d'),
-		sim_homeassistant('\uec6e'),
-		sim_homeassistantcommunitystore('\uec6f'),
-		sim_homebrew('\uec70'),
-		sim_homebridge('\uec71'),
-		sim_homify('\uec72'),
-		sim_honda('\uec73'),
-		sim_hootsuite('\uec74'),
-		sim_hoppscotch('\uec75'),
-		sim_hotelsdotcom('\uec76'),
-		sim_hotjar('\uec77'),
-		sim_houdini('\uec78'),
-		sim_houzz('\uec79'),
-		sim_hp('\uec7a'),
-		sim_html5('\uec7b'),
-		sim_htmlacademy('\uec7c'),
-		sim_httpie('\uec7d'),
-		sim_huawei('\uec7e'),
-		sim_hubspot('\uec7f'),
-		sim_hugo('\uec80'),
-		sim_hulu('\uec81'),
-		sim_humblebundle('\uec82'),
-		sim_hungryjacks('\uec83'),
-		sim_hurriyetemlak('\uec84'),
-		sim_husqvarna('\uec85'),
-		sim_hyper('\uec86'),
-		sim_hyperledger('\uec87'),
-		sim_hypothesis('\uec88'),
-		sim_hyundai('\uec89'),
-		sim_i18next('\uec8a'),
-		sim_iata('\uec8b'),
-		sim_ibeacon('\uec8c'),
-		sim_ibm('\uec8d'),
-		sim_ibmcloud('\uec8e'),
-		sim_ibmwatson('\uec8f'),
-		sim_iceland('\uec90'),
-		sim_icinga('\uec91'),
-		sim_icloud('\uec92'),
-		sim_icomoon('\uec93'),
-		sim_icon('\uec94'),
-		sim_iconfinder('\uec95'),
-		sim_iconify('\uec96'),
-		sim_iconjar('\uec97'),
-		sim_icons8('\uec98'),
-		sim_icq('\uec99'),
-		sim_ieee('\uec9a'),
-		sim_ifixit('\uec9b'),
-		sim_ifood('\uec9c'),
-		sim_ifttt('\uec9d'),
-		sim_iheartradio('\uec9e'),
-		sim_ikea('\uec9f'),
-		sim_imagej('\ueca0'),
-		sim_imdb('\ueca1'),
-		sim_imgur('\ueca2'),
-		sim_immer('\ueca3'),
-		sim_imou('\ueca4'),
-		sim_indeed('\ueca5'),
-		sim_infiniti('\ueca6'),
-		sim_influxdb('\ueca7'),
-		sim_informatica('\ueca8'),
-		sim_infosys('\ueca9'),
-		sim_ingress('\uecaa'),
-		sim_inkscape('\uecab'),
-		sim_insomnia('\uecac'),
-		sim_instacart('\uecad'),
-		sim_instagram('\uecae'),
-		sim_instapaper('\uecaf'),
-		sim_instatus('\uecb0'),
-		sim_instructables('\uecb1'),
-		sim_integromat('\uecb2'),
-		sim_intel('\uecb3'),
-		sim_intellijidea('\uecb4'),
-		sim_interactjs('\uecb5'),
-		sim_intercom('\uecb6'),
-		sim_intermarche('\uecb7'),
-		sim_internetarchive('\uecb8'),
-		sim_internetexplorer('\uecb9'),
-		sim_intigriti('\uecba'),
-		sim_invision('\uecbb'),
-		sim_invoiceninja('\uecbc'),
-		sim_iobroker('\uecbd'),
-		sim_ionic('\uecbe'),
-		sim_ionos('\uecbf'),
-		sim_ios('\uecc0'),
-		sim_iota('\uecc1'),
-		sim_ipfs('\uecc2'),
-		sim_issuu('\uecc3'),
-		sim_istio('\uecc4'),
-		sim_itchdotio('\uecc5'),
-		sim_iterm2('\uecc6'),
-		sim_itunes('\uecc7'),
-		sim_iveco('\uecc8'),
-		sim_jabber('\uecc9'),
-		sim_jaguar('\uecca'),
-		sim_jamboard('\ueccb'),
-		sim_jameson('\ueccc'),
-		sim_jamstack('\ueccd'),
-		sim_jasmine('\uecce'),
-		sim_javascript('\ueccf'),
-		sim_jbl('\uecd0'),
-		sim_jcb('\uecd1'),
-		sim_jeep('\uecd2'),
-		sim_jekyll('\uecd3'),
-		sim_jellyfin('\uecd4'),
-		sim_jenkins('\uecd5'),
-		sim_jenkinsx('\uecd6'),
-		sim_jest('\uecd7'),
-		sim_jet('\uecd8'),
-		sim_jetbrains('\uecd9'),
-		sim_jetpackcompose('\uecda'),
-		sim_jfrog('\uecdb'),
-		sim_jfrogbintray('\uecdc'),
-		sim_jinja('\uecdd'),
-		sim_jira('\uecde'),
-		sim_jirasoftware('\uecdf'),
-		sim_jitsi('\uece0'),
-		sim_johndeere('\uece1'),
-		sim_joomla('\uece2'),
-		sim_joplin('\uece3'),
-		sim_jordan('\uece4'),
-		sim_jpeg('\uece5'),
-		sim_jquery('\uece6'),
-		sim_jrgroup('\uece7'),
-		sim_jsdelivr('\uece8'),
-		sim_jsfiddle('\uece9'),
-		sim_json('\uecea'),
-		sim_jsonwebtokens('\ueceb'),
-		sim_jss('\uecec'),
-		sim_julia('\ueced'),
-		sim_junipernetworks('\uecee'),
-		sim_junit5('\uecef'),
-		sim_jupyter('\uecf0'),
-		sim_justeat('\uecf1'),
-		sim_justgiving('\uecf2'),
-		sim_k3s('\uecf3'),
-		sim_kaggle('\uecf4'),
-		sim_kahoot('\uecf5'),
-		sim_kaios('\uecf6'),
-		sim_kakao('\uecf7'),
-		sim_kakaotalk('\uecf8'),
-		sim_kalilinux('\uecf9'),
-		sim_kaniko('\uecfa'),
-		sim_karlsruherverkehrsverbund('\uecfb'),
-		sim_kasasmart('\uecfc'),
-		sim_kashflow('\uecfd'),
-		sim_kaspersky('\uecfe'),
-		sim_katacoda('\uecff'),
-		sim_katana('\ued00'),
-		sim_kaufland('\ued01'),
-		sim_kde('\ued02'),
-		sim_kdenlive('\ued03'),
-		sim_keepassxc('\ued04'),
-		sim_kentico('\ued05'),
-		sim_keras('\ued06'),
-		sim_keybase('\ued07'),
-		sim_keycdn('\ued08'),
-		sim_kfc('\ued09'),
-		sim_khanacademy('\ued0a'),
-		sim_khronosgroup('\ued0b'),
-		sim_kia('\ued0c'),
-		sim_kibana('\ued0d'),
-		sim_kickstarter('\ued0e'),
-		sim_kik('\ued0f'),
-		sim_kingstontechnology('\ued10'),
-		sim_kinopoisk('\ued11'),
-		sim_kirby('\ued12'),
-		sim_kitsu('\ued13'),
-		sim_klarna('\ued14'),
-		sim_klm('\ued15'),
-		sim_klook('\ued16'),
-		sim_knative('\ued17'),
-		sim_knowledgebase('\ued18'),
-		sim_known('\ued19'),
-		sim_koa('\ued1a'),
-		sim_koc('\ued1b'),
-		sim_kodi('\ued1c'),
-		sim_kofax('\ued1d'),
-		sim_kofi('\ued1e'),
-		sim_komoot('\ued1f'),
-		sim_konami('\ued20'),
-		sim_kong('\ued21'),
-		sim_kongregate('\ued22'),
-		sim_konva('\ued23'),
-		sim_kotlin('\ued24'),
-		sim_krita('\ued25'),
-		sim_ktm('\ued26'),
-		sim_kuaishou('\ued27'),
-		sim_kubernetes('\ued28'),
-		sim_kubuntu('\ued29'),
-		sim_kuma('\ued2a'),
-		sim_kyocera('\ued2b'),
-		sim_labview('\ued2c'),
-		sim_lada('\ued2d'),
-		sim_lamborghini('\ued2e'),
-		sim_landrover('\ued2f'),
-		sim_lapce('\ued30'),
-		sim_laragon('\ued31'),
-		sim_laravel('\ued32'),
-		sim_laravelhorizon('\ued33'),
-		sim_laravelnova('\ued34'),
-		sim_lastdotfm('\ued35'),
-		sim_lastpass('\ued36'),
-		sim_latex('\ued37'),
-		sim_launchpad('\ued38'),
-		sim_lazarus('\ued39'),
-		sim_lbry('\ued3a'),
-		sim_leaderprice('\ued3b'),
-		sim_leaflet('\ued3c'),
-		sim_leanpub('\ued3d'),
-		sim_leetcode('\ued3e'),
-		sim_lemmy('\ued3f'),
-		sim_lenovo('\ued40'),
-		sim_lens('\ued41'),
-		sim_lerna('\ued42'),
-		sim_leroymerlin('\ued43'),
-		sim_less('\ued44'),
-		sim_letsencrypt('\ued45'),
-		sim_letterboxd('\ued46'),
-		sim_levelsdotfyi('\ued47'),
-		sim_lg('\ued48'),
-		sim_lgtm('\ued49'),
-		sim_liberapay('\ued4a'),
-		sim_librariesdotio('\ued4b'),
-		sim_librarything('\ued4c'),
-		sim_libreoffice('\ued4d'),
-		sim_libuv('\ued4e'),
-		sim_lichess('\ued4f'),
-		sim_lidl('\ued50'),
-		sim_lifx('\ued51'),
-		sim_lighthouse('\ued52'),
-		sim_line('\ued53'),
-		sim_lineageos('\ued54'),
-		sim_linear('\ued55'),
-		sim_linkedin('\ued56'),
-		sim_linkerd('\ued57'),
-		sim_linkfire('\ued58'),
-		sim_linktree('\ued59'),
-		sim_linode('\ued5a'),
-		sim_linux('\ued5b'),
-		sim_linuxcontainers('\ued5c'),
-		sim_linuxfoundation('\ued5d'),
-		sim_linuxmint('\ued5e'),
-		sim_lionair('\ued5f'),
-		sim_liquibase('\ued60'),
-		sim_lit('\ued61'),
-		sim_litecoin('\ued62'),
-		sim_livechat('\ued63'),
-		sim_livejournal('\ued64'),
-		sim_livewire('\ued65'),
-		sim_llvm('\ued66'),
-		sim_lmms('\ued67'),
-		sim_lodash('\ued68'),
-		sim_logitech('\ued69'),
-		sim_logmein('\ued6a'),
-		sim_logstash('\ued6b'),
-		sim_looker('\ued6c'),
-		sim_loom('\ued6d'),
-		sim_loop('\ued6e'),
-		sim_loopback('\ued6f'),
-		sim_lospec('\ued70'),
-		sim_lotpolishairlines('\ued71'),
-		sim_lua('\ued72'),
-		sim_lubuntu('\ued73'),
-		sim_lufthansa('\ued74'),
-		sim_lumen('\ued75'),
-		sim_lunacy('\ued76'),
-		sim_lydia('\ued77'),
-		sim_lyft('\ued78'),
-		sim_m('\ue901'),
-		sim_maas('\ued79'),
-		sim_macos('\ued7a'),
-		sim_macys('\ued7b'),
-		sim_magasinsu('\ued7c'),
-		sim_magento('\ued7d'),
-		sim_magisk('\ued7e'),
-		sim_mailchimp('\ued7f'),
-		sim_maildotru('\ued80'),
-		sim_mailgun('\ued81'),
-		sim_majorleaguehacking('\ued82'),
-		sim_makerbot('\ued83'),
-		sim_mamp('\ued84'),
-		sim_man('\ued85'),
-		sim_manageiq('\ued86'),
-		sim_manjaro('\ued87'),
-		sim_mapbox('\ued88'),
-		sim_mariadb('\ued89'),
-		sim_mariadbfoundation('\ued8a'),
-		sim_markdown('\ued8b'),
-		sim_marketo('\ued8c'),
-		sim_marko('\ued8d'),
-		sim_marriott('\ued8e'),
-		sim_maserati('\ued8f'),
-		sim_mastercard('\ued90'),
-		sim_mastercomfig('\ued91'),
-		sim_mastodon('\ued92'),
-		sim_materialdesign('\ued93'),
-		sim_materialdesignicons('\ued94'),
-		sim_matomo('\ued95'),
-		sim_matrix('\ued96'),
-		sim_matterdotjs('\ued97'),
-		sim_mattermost('\ued98'),
-		sim_matternet('\ued99'),
-		sim_max('\ued9a'),
-		sim_maxplanckgesellschaft('\ued9b'),
-		sim_maytag('\ued9c'),
-		sim_mazda('\ued9d'),
-		sim_mcafee('\ued9e'),
-		sim_mcdonalds('\ued9f'),
-		sim_mclaren('\ueda0'),
-		sim_mdbook('\ueda1'),
-		sim_mdnwebdocs('\ueda2'),
-		sim_mdx('\ueda3'),
-		sim_mediafire('\ueda4'),
-		sim_mediamarkt('\ueda5'),
-		sim_mediatek('\ueda6'),
-		sim_mediatemple('\ueda7'),
-		sim_medium('\ueda8'),
-		sim_meetup('\ueda9'),
-		sim_mega('\uedaa'),
-		sim_mendeley('\uedab'),
-		sim_mercedes('\uedac'),
-		sim_merck('\uedad'),
-		sim_mercurial('\uedae'),
-		sim_messenger('\uedaf'),
-		sim_meta('\uedb0'),
-		sim_metabase('\uedb1'),
-		sim_metafilter('\uedb2'),
-		sim_meteor('\uedb3'),
-		sim_metro('\uedb4'),
-		sim_metrodelaciudaddemexico('\uedb5'),
-		sim_metrodemadrid('\uedb6'),
-		sim_metrodeparis('\uedb7'),
-		sim_mewe('\uedb8'),
-		sim_microbit('\uedb9'),
-		sim_microdotblog('\uedba'),
-		sim_microgenetics('\uedbb'),
-		sim_micropython('\uedbc'),
-		sim_microsoft('\uedbd'),
-		sim_microsoftacademic('\uedbe'),
-		sim_microsoftaccess('\uedbf'),
-		sim_microsoftazure('\uedc0'),
-		sim_microsoftbing('\uedc1'),
-		sim_microsoftedge('\uedc2'),
-		sim_microsoftexcel('\uedc3'),
-		sim_microsoftexchange('\uedc4'),
-		sim_microsoftoffice('\uedc5'),
-		sim_microsoftonedrive('\uedc6'),
-		sim_microsoftonenote('\uedc7'),
-		sim_microsoftoutlook('\uedc8'),
-		sim_microsoftpowerpoint('\uedc9'),
-		sim_microsoftsharepoint('\uedca'),
-		sim_microsoftsqlserver('\uedcb'),
-		sim_microsoftteams('\uedcc'),
-		sim_microsofttranslator('\uedcd'),
-		sim_microsoftvisio('\uedce'),
-		sim_microsoftword('\uedcf'),
-		sim_microstrategy('\uedd0'),
-		sim_midi('\uedd1'),
-		sim_minds('\uedd2'),
-		sim_minecraft('\uedd3'),
-		sim_minetest('\uedd4'),
-		sim_mini('\uedd5'),
-		sim_minutemailer('\uedd6'),
-		sim_miro('\uedd7'),
-		sim_mitsubishi('\uedd8'),
-		sim_mix('\uedd9'),
-		sim_mixcloud('\uedda'),
-		sim_mlb('\ueddb'),
-		sim_mlflow('\ueddc'),
-		sim_mobx('\ueddd'),
-		sim_mobxstatetree('\uedde'),
-		sim_mocha('\ueddf'),
-		sim_modx('\uede0'),
-		sim_mojangstudios('\uede1'),
-		sim_moleculer('\uede2'),
-		sim_momenteo('\uede3'),
-		sim_monero('\uede4'),
-		sim_moneygram('\uede5'),
-		sim_mongodb('\uede6'),
-		sim_monkeytie('\uede7'),
-		sim_monoprix('\uede8'),
-		sim_monster('\uede9'),
-		sim_monzo('\uedea'),
-		sim_moo('\uedeb'),
-		sim_morrisons('\uedec'),
-		sim_moscowmetro('\ueded'),
-		sim_motorola('\uedee'),
-		sim_mozilla('\uedef'),
-		sim_msi('\uedf0'),
-		sim_mta('\uedf1'),
-		sim_mtr('\uedf2'),
-		sim_mui('\uedf3'),
-		sim_mulesoft('\uedf4'),
-		sim_muller('\uedf5'),
-		sim_mumble('\uedf6'),
-		sim_musescore('\uedf7'),
-		sim_musicbrainz('\uedf8'),
-		sim_mxlinux('\uedf9'),
-		sim_myanimelist('\uedfa'),
-		sim_myob('\uedfb'),
-		sim_myspace('\uedfc'),
-		sim_mysql('\uedfd'),
-		sim_n26('\uedfe'),
-		sim_namebase('\uedff'),
-		sim_namecheap('\uee00'),
-		sim_nano('\uee01'),
-		sim_nasa('\uee02'),
-		sim_nationalgrid('\uee03'),
-		sim_nativescript('\uee04'),
-		sim_naver('\uee05'),
-		sim_nba('\uee06'),
-		sim_nbb('\uee07'),
-		sim_ndr('\uee08'),
-		sim_nec('\uee09'),
-		sim_neo4j('\uee0a'),
-		sim_neovim('\uee0b'),
-		sim_nestjs('\uee0c'),
-		sim_netapp('\uee0d'),
-		sim_netbsd('\uee0e'),
-		sim_netflix('\uee0f'),
-		sim_netlify('\uee10'),
-		sim_nette('\uee11'),
-		sim_netto('\uee12'),
-		sim_newbalance('\uee13'),
-		sim_newjapanprowrestling('\uee14'),
-		sim_newrelic('\uee15'),
-		sim_newyorktimes('\uee16'),
-		sim_nextcloud('\uee17'),
-		sim_nextdoor('\uee18'),
-		sim_nextdotjs('\uee19'),
-		sim_nfc('\uee1a'),
-		sim_nginx('\uee1b'),
-		sim_ngrok('\uee1c'),
-		sim_niconico('\uee1d'),
-		sim_nike('\uee1e'),
-		sim_nim('\uee1f'),
-		sim_nintendo('\uee20'),
-		sim_nintendo3ds('\uee21'),
-		sim_nintendogamecube('\uee22'),
-		sim_nintendonetwork('\uee23'),
-		sim_nintendoswitch('\uee24'),
-		sim_nissan('\uee25'),
-		sim_nixos('\uee26'),
-		sim_nodedotjs('\uee27'),
-		sim_nodemon('\uee28'),
-		sim_nodered('\uee29'),
-		sim_nokia('\uee2a'),
-		sim_norco('\uee2b'),
-		sim_nordvpn('\uee2c'),
-		sim_norwegian('\uee2d'),
-		sim_notepadplusplus('\uee2e'),
-		sim_notion('\uee2f'),
-		sim_notist('\uee30'),
-		sim_nounproject('\uee31'),
-		sim_now('\uee32'),
-		sim_npm('\uee33'),
-		sim_nrwl('\uee34'),
-		sim_nubank('\uee35'),
-		sim_nucleo('\uee36'),
-		sim_nuget('\uee37'),
-		sim_nuke('\uee38'),
-		sim_numba('\uee39'),
-		sim_numpy('\uee3a'),
-		sim_nutanix('\uee3b'),
-		sim_nuxtdotjs('\uee3c'),
-		sim_nvidia('\uee3d'),
-		sim_nx('\uee3e'),
-		sim_nzxt('\uee3f'),
-		sim_observable('\uee40'),
-		sim_obsidian('\uee41'),
-		sim_obsstudio('\uee42'),
-		sim_ocaml('\uee43'),
-		sim_octanerender('\uee44'),
-		sim_octave('\uee45'),
-		sim_octopusdeploy('\uee46'),
-		sim_oculus('\uee47'),
-		sim_odnoklassniki('\uee48'),
-		sim_odysee('\uee49'),
-		sim_ohdear('\uee4a'),
-		sim_okcupid('\uee4b'),
-		sim_okta('\uee4c'),
-		sim_oneplus('\uee4d'),
-		sim_onlyfans('\uee4e'),
-		sim_onlyoffice('\uee4f'),
-		sim_onnx('\uee50'),
-		sim_onstar('\uee51'),
-		sim_opel('\uee52'),
-		sim_openaccess('\uee53'),
-		sim_openai('\uee54'),
-		sim_openaigym('\uee55'),
-		sim_openapiinitiative('\uee56'),
-		sim_openbadges('\uee57'),
-		sim_openbsd('\uee58'),
-		sim_openbugbounty('\uee59'),
-		sim_opencollective('\uee5a'),
-		sim_opencontainersinitiative('\uee5b'),
-		sim_opencv('\uee5c'),
-		sim_openfaas('\uee5d'),
-		sim_opengl('\uee5e'),
-		sim_openid('\uee5f'),
-		sim_openjdk('\uee60'),
-		sim_openlayers('\uee61'),
-		sim_openmined('\uee62'),
-		sim_opennebula('\uee63'),
-		sim_opensea('\uee64'),
-		sim_opensearch('\uee65'),
-		sim_opensourceinitiative('\uee66'),
-		sim_openssl('\uee67'),
-		sim_openstack('\uee68'),
-		sim_openstreetmap('\uee69'),
-		sim_opensuse('\uee6a'),
-		sim_opentelemetry('\uee6b'),
-		sim_openvpn('\uee6c'),
-		sim_openwrt('\uee6d'),
-		sim_openzfs('\uee6e'),
-		sim_opera('\uee6f'),
-		sim_opnsense('\uee70'),
-		sim_opsgenie('\uee71'),
-		sim_opslevel('\uee72'),
-		sim_oracle('\uee73'),
-		sim_orcid('\uee74'),
-		sim_oreilly('\uee75'),
-		sim_org('\uee76'),
-		sim_origin('\uee77'),
-		sim_osano('\uee78'),
-		sim_oshkosh('\uee79'),
-		sim_osmc('\uee7a'),
-		sim_osu('\uee7b'),
-		sim_otto('\uee7c'),
-		sim_overcast('\uee7d'),
-		sim_overleaf('\uee7e'),
-		sim_ovh('\uee7f'),
-		sim_owasp('\uee80'),
-		sim_oxygen('\uee81'),
-		sim_oyo('\uee82'),
-		sim_p5dotjs('\uee83'),
-		sim_packagist('\uee84'),
-		sim_packer('\uee85'),
-		sim_paddypower('\uee86'),
-		sim_pagekit('\uee87'),
-		sim_pagerduty('\uee88'),
-		sim_pagespeedinsights('\uee89'),
-		sim_pagseguro('\uee8a'),
-		sim_palantir('\uee8b'),
-		sim_paloaltosoftware('\uee8c'),
-		sim_pandas('\uee8d'),
-		sim_pandora('\uee8e'),
-		sim_pantheon('\uee8f'),
-		sim_paperspace('\uee90'),
-		sim_paritysubstrate('\uee91'),
-		sim_parsedotly('\uee92'),
-		sim_passport('\uee93'),
-		sim_password('\ue900'),
-		sim_pastebin('\uee94'),
-		sim_patreon('\uee95'),
-		sim_payoneer('\uee96'),
-		sim_paypal('\uee97'),
-		sim_paytm('\uee98'),
-		sim_pcgamingwiki('\uee99'),
-		sim_peakdesign('\uee9a'),
-		sim_peertube('\uee9b'),
-		sim_pegasusairlines('\uee9c'),
-		sim_pelican('\uee9d'),
-		sim_peloton('\uee9e'),
-		sim_penny('\uee9f'),
-		sim_pepsi('\ueea0'),
-		sim_percy('\ueea1'),
-		sim_perforce('\ueea2'),
-		sim_perl('\ueea3'),
-		sim_personio('\ueea4'),
-		sim_petsathome('\ueea5'),
-		sim_peugeot('\ueea6'),
-		sim_pexels('\ueea7'),
-		sim_pfsense('\ueea8'),
-		sim_phabricator('\ueea9'),
-		sim_philipshue('\ueeaa'),
-		sim_phonepe('\ueeab'),
-		sim_photobucket('\ueeac'),
-		sim_photocrowd('\ueead'),
-		sim_photopea('\ueeae'),
-		sim_php('\ueeaf'),
-		sim_phpmyadmin('\ueeb0'),
-		sim_phpstorm('\ueeb1'),
-		sim_picardsurgeles('\ueeb2'),
-		sim_picartodottv('\ueeb3'),
-		sim_picnic('\ueeb4'),
-		sim_picpay('\ueeb5'),
-		sim_pihole('\ueeb6'),
-		sim_pimcore('\ueeb7'),
-		sim_pinboard('\ueeb8'),
-		sim_pingdom('\ueeb9'),
-		sim_pinterest('\ueeba'),
-		sim_pioneerdj('\ueebb'),
-		sim_pivotaltracker('\ueebc'),
-		sim_piwigo('\ueebd'),
-		sim_pix('\ueebe'),
-		sim_pixabay('\ueebf'),
-		sim_pixiv('\ueec0'),
-		sim_pkgsrc('\ueec1'),
-		sim_planet('\ueec2'),
-		sim_planetscale('\ueec3'),
-		sim_plangrid('\ueec4'),
-		sim_platformdotsh('\ueec5'),
-		sim_platzi('\ueec6'),
-		sim_plausibleanalytics('\ueec7'),
-		sim_playcanvas('\ueec8'),
-		sim_playerdotme('\ueec9'),
-		sim_playerfm('\ueeca'),
-		sim_playstation('\ueecb'),
-		sim_playstation2('\ueecc'),
-		sim_playstation3('\ueecd'),
-		sim_playstation4('\ueece'),
-		sim_playstation5('\ueecf'),
-		sim_playstationvita('\ueed0'),
-		sim_pleroma('\ueed1'),
-		sim_plesk('\ueed2'),
-		sim_plex('\ueed3'),
-		sim_plotly('\ueed4'),
-		sim_pluralsight('\ueed5'),
-		sim_plurk('\ueed6'),
-		sim_pluscodes('\ueed7'),
-		sim_pm2('\ueed8'),
-		sim_pnpm('\ueed9'),
-		sim_pocket('\ueeda'),
-		sim_pocketcasts('\ueedb'),
-		sim_podcastaddict('\ueedc'),
-		sim_podman('\ueedd'),
-		sim_poetry('\ueede'),
-		sim_pointy('\ueedf'),
-		sim_pokemon('\ueee0'),
-		sim_polkadot('\ueee1'),
-		sim_poly('\ueee2'),
-		sim_polymerproject('\ueee3'),
-		sim_polywork('\ueee4'),
-		sim_popos('\ueee5'),
-		sim_porsche('\ueee6'),
-		sim_portainer('\ueee7'),
-		sim_postcss('\ueee8'),
-		sim_postgresql('\ueee9'),
-		sim_postman('\ueeea'),
-		sim_postmates('\ueeeb'),
-		sim_powerapps('\ueeec'),
-		sim_powerautomate('\ueeed'),
-		sim_powerbi('\ueeee'),
-		sim_powers('\ueeef'),
-		sim_powershell('\ueef0'),
-		sim_powervirtualagents('\ueef1'),
-		sim_prdotco('\ueef2'),
-		sim_preact('\ueef3'),
-		sim_precommit('\ueef4'),
-		sim_premierleague('\ueef5'),
-		sim_prestashop('\ueef6'),
-		sim_presto('\ueef7'),
-		sim_prettier('\ueef8'),
-		sim_prezi('\ueef9'),
-		sim_prime('\ueefa'),
-		sim_primevideo('\ueefb'),
-		sim_prisma('\ueefc'),
-		sim_prismic('\ueefd'),
-		sim_privateinternetaccess('\ueefe'),
-		sim_probot('\ueeff'),
-		sim_processingfoundation('\uef00'),
-		sim_processwire('\uef01'),
-		sim_producthunt('\uef02'),
-		sim_progate('\uef03'),
-		sim_progress('\uef04'),
-		sim_prometheus('\uef05'),
-		sim_prosieben('\uef06'),
-		sim_protocolsdotio('\uef07'),
-		sim_protodotio('\uef08'),
-		sim_protondb('\uef09'),
-		sim_protonmail('\uef0a'),
-		sim_protonvpn('\uef0b'),
-		sim_protools('\uef0c'),
-		sim_protractor('\uef0d'),
-		sim_proxmox('\uef0e'),
-		sim_pubg('\uef0f'),
-		sim_publons('\uef10'),
-		sim_pubmed('\uef11'),
-		sim_pug('\uef12'),
-		sim_pulumi('\uef13'),
-		sim_puma('\uef14'),
-		sim_puppet('\uef15'),
-		sim_puppeteer('\uef16'),
-		sim_purescript('\uef17'),
-		sim_purgecss('\uef18'),
-		sim_purism('\uef19'),
-		sim_pusher('\uef1a'),
-		sim_pwa('\uef1b'),
-		sim_px('\ue905'),
-		sim_pycharm('\uef1c'),
-		sim_pypi('\uef1d'),
-		sim_pypy('\uef1e'),
-		sim_pyscaffold('\uef1f'),
-		sim_pytest('\uef20'),
-		sim_python('\uef21'),
-		sim_pytorch('\uef22'),
-		sim_pytorchlightning('\uef23'),
-		sim_pyup('\uef24'),
-		sim_qantas('\uef25'),
-		sim_qatarairways('\uef26'),
-		sim_qemu('\uef27'),
-		sim_qgis('\uef28'),
-		sim_qi('\uef29'),
-		sim_qiita('\uef2a'),
-		sim_qiskit('\uef2b'),
-		sim_qiwi('\uef2c'),
-		sim_qt('\uef2d'),
-		sim_qualcomm('\uef2e'),
-		sim_qualtrics('\uef2f'),
-		sim_qualys('\uef30'),
-		sim_quantcast('\uef31'),
-		sim_quantconnect('\uef32'),
-		sim_quarkus('\uef33'),
-		sim_quasar('\uef34'),
-		sim_qubesos('\uef35'),
-		sim_quest('\uef36'),
-		sim_quickbooks('\uef37'),
-		sim_quicklook('\uef38'),
-		sim_quicktime('\uef39'),
-		sim_quip('\uef3a'),
-		sim_quora('\uef3b'),
-		sim_qwiklabs('\uef3c'),
-		sim_qzone('\uef3d'),
-		sim_r('\uef3e'),
-		sim_r3('\uef3f'),
-		sim_rabbitmq('\uef40'),
-		sim_racket('\uef41'),
-		sim_radar('\uef42'),
-		sim_radiopublic('\uef43'),
-		sim_railway('\uef44'),
-		sim_rainmeter('\uef45'),
-		sim_rakuten('\uef46'),
-		sim_ram('\uef47'),
-		sim_rancher('\uef48'),
-		sim_rarible('\uef49'),
-		sim_rasa('\uef4a'),
-		sim_raspberrypi('\uef4b'),
-		sim_ray('\uef4c'),
-		sim_razer('\uef4d'),
-		sim_razorpay('\uef4e'),
-		sim_react('\uef4f'),
-		sim_reactivex('\uef50'),
-		sim_reactos('\uef51'),
-		sim_reactquery('\uef52'),
-		sim_reactrouter('\uef53'),
-		sim_reacttable('\uef54'),
-		sim_readthedocs('\uef55'),
-		sim_realm('\uef56'),
-		sim_reason('\uef57'),
-		sim_reasonstudios('\uef58'),
-		sim_redbubble('\uef59'),
-		sim_reddit('\uef5a'),
-		sim_redhat('\uef5b'),
-		sim_redhatopenshift('\uef5c'),
-		sim_redis('\uef5d'),
-		sim_redmine('\uef5e'),
-		sim_redux('\uef5f'),
-		sim_reduxsaga('\uef60'),
-		sim_redwoodjs('\uef61'),
-		sim_reebok('\uef62'),
-		sim_relianceindustrieslimited('\uef63'),
-		sim_remix('\uef64'),
-		sim_renault('\uef65'),
-		sim_renovatebot('\uef66'),
-		sim_renpy('\uef67'),
-		sim_renren('\uef68'),
-		sim_replit('\uef69'),
-		sim_republicofgamers('\uef6a'),
-		sim_rescript('\uef6b'),
-		sim_rescuetime('\uef6c'),
-		sim_researchgate('\uef6d'),
-		sim_resurrectionremixos('\uef6e'),
-		sim_retroarch('\uef6f'),
-		sim_retropie('\uef70'),
-		sim_revealdotjs('\uef71'),
-		sim_reverbnation('\uef72'),
-		sim_revolut('\uef73'),
-		sim_revue('\uef74'),
-		sim_rewe('\uef75'),
-		sim_rezgo('\uef76'),
-		sim_rhinoceros('\uef77'),
-		sim_rider('\uef78'),
-		sim_rimacautomobili('\uef79'),
-		sim_ring('\uef7a'),
-		sim_riotgames('\uef7b'),
-		sim_ripple('\uef7c'),
-		sim_riseup('\uef7d'),
-		sim_roamresearch('\uef7e'),
-		sim_roblox('\uef7f'),
-		sim_robotframework('\uef80'),
-		sim_rocketdotchat('\uef81'),
-		sim_rocksdb('\uef82'),
-		sim_rockylinux('\uef83'),
-		sim_roku('\uef84'),
-		sim_rollsroyce('\uef85'),
-		sim_rollupdotjs('\uef86'),
-		sim_rome('\uef87'),
-		sim_roots('\uef88'),
-		sim_rootsbedrock('\uef89'),
-		sim_rootssage('\uef8a'),
-		sim_ros('\uef8b'),
-		sim_rossmann('\uef8c'),
-		sim_rotaryinternational('\uef8d'),
-		sim_rottentomatoes('\uef8e'),
-		sim_roundcube('\uef8f'),
-		sim_rss('\uef90'),
-		sim_rstudio('\uef91'),
-		sim_rte('\uef92'),
-		sim_rtl('\uef93'),
-		sim_rtlzwei('\uef94'),
-		sim_rubocop('\uef95'),
-		sim_ruby('\uef96'),
-		sim_rubygems('\uef97'),
-		sim_rubyonrails('\uef98'),
-		sim_rubysinatra('\uef99'),
-		sim_runkeeper('\uef9a'),
-		sim_runkit('\uef9b'),
-		sim_rust('\uef9c'),
-		sim_rxdb('\uef9d'),
-		sim_ryanair('\uef9e'),
-		sim_s7airlines('\uef9f'),
-		sim_sabanci('\uefa0'),
-		sim_safari('\uefa1'),
-		sim_sahibinden('\uefa2'),
-		sim_sailfishos('\uefa3'),
-		sim_salesforce('\uefa4'),
-		sim_saltproject('\uefa5'),
-		sim_samsung('\uefa6'),
-		sim_samsungpay('\uefa7'),
-		sim_sandisk('\uefa8'),
-		sim_sanfranciscomunicipalrailway('\uefa9'),
-		sim_saopaulometro('\uefaa'),
-		sim_sap('\uefab'),
-		sim_sass('\uefac'),
-		sim_sat1('\uefad'),
-		sim_saturn('\uefae'),
-		sim_saucelabs('\uefaf'),
-		sim_scala('\uefb0'),
-		sim_scaleway('\uefb1'),
-		sim_scania('\uefb2'),
-		sim_schneiderelectric('\uefb3'),
-		sim_scikitlearn('\uefb4'),
-		sim_scipy('\uefb5'),
-		sim_scopus('\uefb6'),
-		sim_scpfoundation('\uefb7'),
-		sim_scratch('\uefb8'),
-		sim_screencastify('\uefb9'),
-		sim_scribd('\uefba'),
-		sim_scrimba('\uefbb'),
-		sim_scrollreveal('\uefbc'),
-		sim_scrumalliance('\uefbd'),
-		sim_scrutinizerci('\uefbe'),
-		sim_seagate('\uefbf'),
-		sim_seat('\uefc0'),
-		sim_sefaria('\uefc1'),
-		sim_sega('\uefc2'),
-		sim_selenium('\uefc3'),
-		sim_sellfy('\uefc4'),
-		sim_semanticrelease('\uefc5'),
-		sim_semanticuireact('\uefc6'),
-		sim_semanticweb('\uefc7'),
-		sim_semaphoreci('\uefc8'),
-		sim_semver('\uefc9'),
-		sim_sencha('\uefca'),
-		sim_sennheiser('\uefcb'),
-		sim_sensu('\uefcc'),
-		sim_sentry('\uefcd'),
-		sim_sepa('\uefce'),
-		sim_sequelize('\uefcf'),
-		sim_serverfault('\uefd0'),
-		sim_serverless('\uefd1'),
-		sim_sfml('\uefd2'),
-		sim_shadow('\uefd3'),
-		sim_shanghaimetro('\uefd4'),
-		sim_sharp('\uefd5'),
-		sim_shazam('\uefd6'),
-		sim_shell('\uefd7'),
-		sim_shelly('\uefd8'),
-		sim_shenzhenmetro('\uefd9'),
-		sim_shieldsdotio('\uefda'),
-		sim_shikimori('\uefdb'),
-		sim_shopify('\uefdc'),
-		sim_shopware('\uefdd'),
-		sim_shotcut('\uefde'),
-		sim_showpad('\uefdf'),
-		sim_showtime('\uefe0'),
-		sim_shutterstock('\uefe1'),
-		sim_siemens('\uefe2'),
-		sim_signal('\uefe3'),
-		sim_simkl('\uefe4'),
-		sim_simpleanalytics('\uefe5'),
-		sim_simpleicons('\uefe6'),
-		sim_simplenote('\uefe7'),
-		sim_sinaweibo('\uefe8'),
-		sim_singlestore('\uefe9'),
-		sim_sitepoint('\uefea'),
-		sim_sketch('\uefeb'),
-		sim_sketchfab('\uefec'),
-		sim_sketchup('\uefed'),
-		sim_skillshare('\uefee'),
-		sim_skoda('\uefef'),
-		sim_sky('\ueff0'),
-		sim_skynet('\ueff1'),
-		sim_skypack('\ueff2'),
-		sim_skype('\ueff3'),
-		sim_skypeforbusiness('\ueff4'),
-		sim_slack('\ueff5'),
-		sim_slackware('\ueff6'),
-		sim_slashdot('\ueff7'),
-		sim_slickpic('\ueff8'),
-		sim_slides('\ueff9'),
-		sim_slideshare('\ueffa'),
-		sim_smart('\ueffb'),
-		sim_smartthings('\ueffc'),
-		sim_smashdotgg('\ueffd'),
-		sim_smashingmagazine('\ueffe'),
-		sim_smrt('\uefff'),
-		sim_smugmug('\uf000'),
-		sim_snapchat('\uf001'),
-		sim_snapcraft('\uf002'),
-		sim_snowflake('\uf003'),
-		sim_snowpack('\uf004'),
-		sim_snyk('\uf005'),
-		sim_socialblade('\uf006'),
-		sim_society6('\uf007'),
-		sim_socketdotio('\uf008'),
-		sim_sogou('\uf009'),
-		sim_solid('\uf00a'),
-		sim_solidity('\uf00b'),
-		sim_sololearn('\uf00c'),
-		sim_solus('\uf00d'),
-		sim_sonarcloud('\uf00e'),
-		sim_sonarlint('\uf00f'),
-		sim_sonarqube('\uf010'),
-		sim_sonarsource('\uf011'),
-		sim_songkick('\uf012'),
-		sim_songoda('\uf013'),
-		sim_sonicwall('\uf014'),
-		sim_sonos('\uf015'),
-		sim_sony('\uf016'),
-		sim_soundcharts('\uf017'),
-		sim_soundcloud('\uf018'),
-		sim_sourceengine('\uf019'),
-		sim_sourceforge('\uf01a'),
-		sim_sourcegraph('\uf01b'),
-		sim_sourcetree('\uf01c'),
-		sim_southwestairlines('\uf01d'),
-		sim_spacemacs('\uf01e'),
-		sim_spacex('\uf01f'),
-		sim_spacy('\uf020'),
-		sim_sparkar('\uf021'),
-		sim_sparkasse('\uf022'),
-		sim_sparkfun('\uf023'),
-		sim_sparkpost('\uf024'),
-		sim_spdx('\uf025'),
-		sim_speakerdeck('\uf026'),
-		sim_spectrum('\uf027'),
-		sim_speedtest('\uf028'),
-		sim_spinnaker('\uf029'),
-		sim_spinrilla('\uf02a'),
-		sim_splunk('\uf02b'),
-		sim_spond('\uf02c'),
-		sim_spotify('\uf02d'),
-		sim_spotlight('\uf02e'),
-		sim_spreadshirt('\uf02f'),
-		sim_spreaker('\uf030'),
-		sim_spring('\uf031'),
-		sim_spring_creators('\uf032'),
-		sim_springboot('\uf033'),
-		sim_springsecurity('\uf034'),
-		sim_spyderide('\uf035'),
-		sim_sqlite('\uf036'),
-		sim_square('\uf037'),
-		sim_squareenix('\uf038'),
-		sim_squarespace('\uf039'),
-		sim_ssrn('\uf03a'),
-		sim_stackbit('\uf03b'),
-		sim_stackedit('\uf03c'),
-		sim_stackexchange('\uf03d'),
-		sim_stackoverflow('\uf03e'),
-		sim_stackpath('\uf03f'),
-		sim_stackshare('\uf040'),
-		sim_stadia('\uf041'),
-		sim_staffbase('\uf042'),
-		sim_starbucks('\uf043'),
-		sim_stardock('\uf044'),
-		sim_starlingbank('\uf045'),
-		sim_starship('\uf046'),
-		sim_startrek('\uf047'),
-		sim_starz('\uf048'),
-		sim_statamic('\uf049'),
-		sim_statuspage('\uf04a'),
-		sim_statuspal('\uf04b'),
-		sim_steam('\uf04c'),
-		sim_steamdb('\uf04d'),
-		sim_steamdeck('\uf04e'),
-		sim_steamworks('\uf04f'),
-		sim_steelseries('\uf050'),
-		sim_steem('\uf051'),
-		sim_steemit('\uf052'),
-		sim_steinberg('\uf053'),
-		sim_stellar('\uf054'),
-		sim_stencyl('\uf055'),
-		sim_stimulus('\uf056'),
-		sim_stitcher('\uf057'),
-		sim_stmicroelectronics('\uf058'),
-		sim_stopstalk('\uf059'),
-		sim_storyblok('\uf05a'),
-		sim_storybook('\uf05b'),
-		sim_strapi('\uf05c'),
-		sim_strava('\uf05d'),
-		sim_streamlit('\uf05e'),
-		sim_stripe('\uf05f'),
-		sim_strongswan('\uf060'),
-		sim_stubhub('\uf061'),
-		sim_styledcomponents('\uf062'),
-		sim_stylelint('\uf063'),
-		sim_styleshare('\uf064'),
-		sim_stylus('\uf065'),
-		sim_subaru('\uf066'),
-		sim_sublimetext('\uf067'),
-		sim_substack('\uf068'),
-		sim_subversion('\uf069'),
-		sim_suckless('\uf06a'),
-		sim_sumologic('\uf06b'),
-		sim_supabase('\uf06c'),
-		sim_supermicro('\uf06d'),
-		sim_superuser('\uf06e'),
-		sim_surveymonkey('\uf06f'),
-		sim_suse('\uf070'),
-		sim_suzuki('\uf071'),
-		sim_svelte('\uf072'),
-		sim_svg('\uf073'),
-		sim_svgo('\uf074'),
-		sim_swagger('\uf075'),
-		sim_swarm('\uf076'),
-		sim_swc('\uf077'),
-		sim_swift('\uf078'),
-		sim_swiggy('\uf079'),
-		sim_swiper('\uf07a'),
-		sim_symantec('\uf07b'),
-		sim_symfony('\uf07c'),
-		sim_symphony('\uf07d'),
-		sim_sympy('\uf07e'),
-		sim_synology('\uf07f'),
-		sim_tableau('\uf080'),
-		sim_tado('\uf081'),
-		sim_tails('\uf082'),
-		sim_tailwindcss('\uf083'),
-		sim_talend('\uf084'),
-		sim_talenthouse('\uf085'),
-		sim_tampermonkey('\uf086'),
-		sim_taobao('\uf087'),
-		sim_tapas('\uf088'),
-		sim_target('\uf089'),
-		sim_task('\uf08a'),
-		sim_tasmota('\uf08b'),
-		sim_tata('\uf08c'),
-		sim_tauri('\uf08d'),
-		sim_taxbuzz('\uf08e'),
-		sim_teamcity('\uf08f'),
-		sim_teamspeak('\uf090'),
-		sim_teamviewer('\uf091'),
-		sim_ted('\uf092'),
-		sim_teespring('\uf093'),
-		sim_tekton('\uf094'),
-		sim_tele5('\uf095'),
-		sim_telegram('\uf096'),
-		sim_telegraph('\uf097'),
-		sim_temporal('\uf098'),
-		sim_tencentqq('\uf099'),
-		sim_tensorflow('\uf09a'),
-		sim_teradata('\uf09b'),
-		sim_teratail('\uf09c'),
-		sim_terraform('\uf09d'),
-		sim_tesco('\uf09e'),
-		sim_tesla('\uf09f'),
-		sim_testcafe('\uf0a0'),
-		sim_testin('\uf0a1'),
-		sim_testinglibrary('\uf0a2'),
-		sim_tether('\uf0a3'),
-		sim_textpattern('\uf0a4'),
-		sim_thealgorithms('\uf0a5'),
-		sim_theconversation('\uf0a6'),
-		sim_theirishtimes('\uf0a7'),
-		sim_themighty('\uf0a8'),
-		sim_themodelsresource('\uf0a9'),
-		sim_themoviedatabase('\uf0aa'),
-		sim_thenorthface('\uf0ab'),
-		sim_theregister('\uf0ac'),
-		sim_thesoundsresource('\uf0ad'),
-		sim_thespritersresource('\uf0ae'),
-		sim_thewashingtonpost('\uf0af'),
-		sim_thingiverse('\uf0b0'),
-		sim_thinkpad('\uf0b1'),
-		sim_threadless('\uf0b2'),
-		sim_threedotjs('\uf0b3'),
-		sim_threema('\uf0b4'),
-		sim_thumbtack('\uf0b5'),
-		sim_thunderbird('\uf0b6'),
-		sim_thymeleaf('\uf0b7'),
-		sim_ticketmaster('\uf0b8'),
-		sim_tidal('\uf0b9'),
-		sim_tide('\uf0ba'),
-		sim_tietoevry('\uf0bb'),
-		sim_tiktok('\uf0bc'),
-		sim_tile('\uf0bd'),
-		sim_timescale('\uf0be'),
-		sim_tinder('\uf0bf'),
-		sim_tinyletter('\uf0c0'),
-		sim_tmobile('\uf0c1'),
-		sim_tmux('\uf0c2'),
-		sim_todoist('\uf0c3'),
-		sim_toggl('\uf0c4'),
-		sim_tokyometro('\uf0c5'),
-		sim_tomorrowland('\uf0c6'),
-		sim_topcoder('\uf0c7'),
-		sim_toptal('\uf0c8'),
-		sim_torbrowser('\uf0c9'),
-		sim_torproject('\uf0ca'),
-		sim_toshiba('\uf0cb'),
-		sim_toyota('\uf0cc'),
-		sim_tplink('\uf0cd'),
-		sim_tqdm('\uf0ce'),
-		sim_tracklists('\ue906'),
-		sim_traefikmesh('\uf0cf'),
-		sim_trainerroad('\uf0d0'),
-		sim_trakt('\uf0d1'),
-		sim_transportforireland('\uf0d2'),
-		sim_transportforlondon('\uf0d3'),
-		sim_travisci('\uf0d4'),
-		sim_treehouse('\uf0d5'),
-		sim_trello('\uf0d6'),
-		sim_trendmicro('\uf0d7'),
-		sim_treyarch('\uf0d8'),
-		sim_triller('\uf0d9'),
-		sim_trino('\uf0da'),
-		sim_tripadvisor('\uf0db'),
-		sim_tripdotcom('\uf0dc'),
-		sim_trove('\uf0dd'),
-		sim_trpc('\uf0de'),
-		sim_truenas('\uf0df'),
-		sim_trulia('\uf0e0'),
-		sim_trustedshops('\uf0e1'),
-		sim_trustpilot('\uf0e2'),
-		sim_tryhackme('\uf0e3'),
-		sim_tryitonline('\uf0e4'),
-		sim_tsnode('\uf0e5'),
-		sim_tubi('\uf0e6'),
-		sim_tui('\uf0e7'),
-		sim_tumblr('\uf0e8'),
-		sim_tunein('\uf0e9'),
-		sim_turborepo('\uf0ea'),
-		sim_turbosquid('\uf0eb'),
-		sim_turkishairlines('\uf0ec'),
-		sim_tutanota('\uf0ed'),
-		sim_tvtime('\uf0ee'),
-		sim_twilio('\uf0ef'),
-		sim_twitch('\uf0f0'),
-		sim_twitter('\uf0f1'),
-		sim_twoo('\uf0f2'),
-		sim_typeform('\uf0f3'),
-		sim_typescript('\uf0f4'),
-		sim_typo3('\uf0f5'),
-		sim_uber('\uf0f6'),
-		sim_ubereats('\uf0f7'),
-		sim_ubiquiti('\uf0f8'),
-		sim_ubisoft('\uf0f9'),
-		sim_ublockorigin('\uf0fa'),
-		sim_ubuntu('\uf0fb'),
-		sim_udacity('\uf0fc'),
-		sim_udemy('\uf0fd'),
-		sim_ufc('\uf0fe'),
-		sim_uikit('\uf0ff'),
-		sim_ulule('\uf100'),
-		sim_umbraco('\uf101'),
-		sim_unacademy('\uf102'),
-		sim_underarmour('\uf103'),
-		sim_underscoredotjs('\uf104'),
-		sim_undertale('\uf105'),
-		sim_unicode('\uf106'),
-		sim_unilever('\uf107'),
-		sim_unitedairlines('\uf108'),
-		sim_unity('\uf109'),
-		sim_unraid('\uf10a'),
-		sim_unrealengine('\uf10b'),
-		sim_unsplash('\uf10c'),
-		sim_untangle('\uf10d'),
-		sim_untappd('\uf10e'),
-		sim_upcloud('\uf10f'),
-		sim_uplabs('\uf110'),
-		sim_uploaded('\uf111'),
-		sim_ups('\uf112'),
-		sim_upstash('\uf113'),
-		sim_uptobox('\uf114'),
-		sim_upwork('\uf115'),
-		sim_usps('\uf116'),
-		sim_v('\uf117'),
-		sim_v8('\uf118'),
-		sim_vaadin('\uf119'),
-		sim_vagrant('\uf11a'),
-		sim_valorant('\uf11b'),
-		sim_valve('\uf11c'),
-		sim_vapor('\uf11d'),
-		sim_vault('\uf11e'),
-		sim_vauxhall('\uf11f'),
-		sim_vbulletin('\uf120'),
-		sim_vectorlogozone('\uf121'),
-		sim_vectorworks('\uf122'),
-		sim_veeam('\uf123'),
-		sim_veepee('\uf124'),
-		sim_velog('\uf125'),
-		sim_venmo('\uf126'),
-		sim_vercel('\uf127'),
-		sim_verdaccio('\uf128'),
-		sim_veritas('\uf129'),
-		sim_verizon('\uf12a'),
-		sim_vexxhost('\uf12b'),
-		sim_vfairs('\uf12c'),
-		sim_viadeo('\uf12d'),
-		sim_viber('\uf12e'),
-		sim_vim('\uf12f'),
-		sim_vimeo('\uf130'),
-		sim_vimeolivestream('\uf131'),
-		sim_virgin('\uf132'),
-		sim_virtualbox('\uf133'),
-		sim_virustotal('\uf134'),
-		sim_visa('\uf135'),
-		sim_visualstudio('\uf136'),
-		sim_visualstudiocode('\uf137'),
-		sim_vite('\uf138'),
-		sim_vitess('\uf139'),
-		sim_vivaldi('\uf13a'),
-		sim_vivino('\uf13b'),
-		sim_vk('\uf13c'),
-		sim_vlcmediaplayer('\uf13d'),
-		sim_vmware('\uf13e'),
-		sim_vodafone('\uf13f'),
-		sim_volkswagen('\uf140'),
-		sim_volvo('\uf141'),
-		sim_vonage('\uf142'),
-		sim_vowpalwabbit('\uf143'),
-		sim_vox('\uf144'),
-		sim_vsco('\uf145'),
-		sim_vtex('\uf146'),
-		sim_vuedotjs('\uf147'),
-		sim_vuetify('\uf148'),
-		sim_vulkan('\uf149'),
-		sim_vultr('\uf14a'),
-		sim_w3c('\uf14b'),
-		sim_wagtail('\uf14c'),
-		sim_wakatime('\uf14d'),
-		sim_walkman('\uf14e'),
-		sim_wallabag('\uf14f'),
-		sim_walmart('\uf150'),
-		sim_wappalyzer('\uf151'),
-		sim_warnerbros('\uf152'),
-		sim_wattpad('\uf153'),
-		sim_waze('\uf154'),
-		sim_wearos('\uf155'),
-		sim_weasyl('\uf156'),
-		sim_web3dotjs('\uf157'),
-		sim_webassembly('\uf158'),
-		sim_webauthn('\uf159'),
-		sim_webcomponentsdotorg('\uf15a'),
-		sim_webdriverio('\uf15b'),
-		sim_webflow('\uf15c'),
-		sim_webgl('\uf15d'),
-		sim_webhint('\uf15e'),
-		sim_weblate('\uf15f'),
-		sim_webmin('\uf160'),
-		sim_webmoney('\uf161'),
-		sim_webpack('\uf162'),
-		sim_webrtc('\uf163'),
-		sim_webstorm('\uf164'),
-		sim_webtoon('\uf165'),
-		sim_wechat('\uf166'),
-		sim_wegame('\uf167'),
-		sim_weightsandbiases('\uf168'),
-		sim_welcometothejungle('\uf169'),
-		sim_wemo('\uf16a'),
-		sim_westerndigital('\uf16b'),
-		sim_wetransfer('\uf16c'),
-		sim_whatsapp('\uf16d'),
-		sim_wheniwork('\uf16e'),
-		sim_whitesource('\uf16f'),
-		sim_wii('\uf170'),
-		sim_wiiu('\uf171'),
-		sim_wikidata('\uf172'),
-		sim_wikidotjs('\uf173'),
-		sim_wikimediacommons('\uf174'),
-		sim_wikipedia('\uf175'),
-		sim_wikiquote('\uf176'),
-		sim_wikivoyage('\uf177'),
-		sim_winamp('\uf178'),
-		sim_windicss('\uf179'),
-		sim_windows('\uf17a'),
-		sim_windows95('\uf17b'),
-		sim_windowsterminal('\uf17c'),
-		sim_windowsxp('\uf17d'),
-		sim_winmate('\uf17e'),
-		sim_wipro('\uf17f'),
-		sim_wire('\uf180'),
-		sim_wireguard('\uf181'),
-		sim_wireshark('\uf182'),
-		sim_wise('\uf183'),
-		sim_wish('\uf184'),
-		sim_wistia('\uf185'),
-		sim_wix('\uf186'),
-		sim_wizzair('\uf187'),
-		sim_wolfram('\uf188'),
-		sim_wolframlanguage('\uf189'),
-		sim_wolframmathematica('\uf18a'),
-		sim_woo('\uf18b'),
-		sim_woocommerce('\uf18c'),
-		sim_wordpress('\uf18d'),
-		sim_workplace('\uf18e'),
-		sim_worldhealthorganization('\uf18f'),
-		sim_wpengine('\uf190'),
-		sim_wprocket('\uf191'),
-		sim_writedotas('\uf192'),
-		sim_wwe('\uf193'),
-		sim_wwise('\uf194'),
-		sim_xamarin('\uf195'),
-		sim_xaml('\uf196'),
-		sim_xampp('\uf197'),
-		sim_xbox('\uf198'),
-		sim_xcode('\uf199'),
-		sim_xdadevelopers('\uf19a'),
-		sim_xdotorg('\uf19b'),
-		sim_xero('\uf19c'),
-		sim_xfce('\uf19d'),
-		sim_xiaomi('\uf19e'),
-		sim_xilinx('\uf19f'),
-		sim_xing('\uf1a0'),
-		sim_xmpp('\uf1a1'),
-		sim_xrp('\uf1a2'),
-		sim_xsplit('\uf1a3'),
-		sim_xstate('\uf1a4'),
-		sim_yahoo('\uf1a5'),
-		sim_yale('\uf1a6'),
-		sim_yamahacorporation('\uf1a7'),
-		sim_yamahamotorcorporation('\uf1a8'),
-		sim_yammer('\uf1a9'),
-		sim_yarn('\uf1aa'),
-		sim_ycombinator('\uf1ab'),
-		sim_yelp('\uf1ac'),
-		sim_yoast('\uf1ad'),
-		sim_yolo('\uf1ae'),
-		sim_yourtraveldottv('\uf1af'),
-		sim_youtube('\uf1b0'),
-		sim_youtubegaming('\uf1b1'),
-		sim_youtubemusic('\uf1b2'),
-		sim_youtubestudio('\uf1b3'),
-		sim_youtubetv('\uf1b4'),
-		sim_yubico('\uf1b5'),
-		sim_zabka('\uf1b6'),
-		sim_zalando('\uf1b7'),
-		sim_zalo('\uf1b8'),
-		sim_zapier('\uf1b9'),
-		sim_zara('\uf1ba'),
-		sim_zazzle('\uf1bb'),
-		sim_zcash('\uf1bc'),
-		sim_zdf('\uf1bd'),
-		sim_zelle('\uf1be'),
-		sim_zend('\uf1bf'),
-		sim_zendesk('\uf1c0'),
-		sim_zendframework('\uf1c1'),
-		sim_zenn('\uf1c2'),
-		sim_zenodo('\uf1c3'),
-		sim_zerodha('\uf1c4'),
-		sim_zeromq('\uf1c5'),
-		sim_zerply('\uf1c6'),
-		sim_zhihu('\uf1c7'),
-		sim_zig('\uf1c8'),
-		sim_zigbee('\uf1c9'),
-		sim_zillow('\uf1ca'),
-		sim_zingat('\uf1cb'),
-		sim_zoho('\uf1cc'),
-		sim_zoiper('\uf1cd'),
-		sim_zomato('\uf1ce'),
-		sim_zoom('\uf1cf'),
-		sim_zorin('\uf1d0'),
-		sim_zotero('\uf1d1'),
-		sim_zulip('\uf1d2'),
-		sim_zwave('\uf1d3'),
-		sim_zyte('\uf1d4');
-
-        override val typeface: ITypeface by lazy { SimpleIcons }
-    }
-}
